1. fix move parameters float issue

+++ c_mainwindow.py
@@ -470,7 +470,7 @@ class MainUi(QMainWindow):
 
         screen = QDesktopWidget().screenGeometry()
         size = self.geometry()
-        self.move((screen.width() - size.width()) / 2, (screen.height() - size.height()) / 2)
+        self.move(int((screen.width() - size.width()) / 2), int((screen.height() - size.height()) / 2))
 
     def play_status_changed(self, changed, status):
         log.debug("changed = %d", changed)
